Fixed linear search in extract prepared transaction.
git-svn-id: https://svn.neo4j.org/components/kernel/branches/ha@5640 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/main/java/org/neo4j/kernel/impl/transaction/xaframework/XaDataSource.java
@@ -346,7 +346,7 @@ public abstract class XaDataSource
         throw new UnsupportedOperationException();
     }
     
-    public ReadableByteChannel getPreparedTransaction( long identifier ) throws IOException
+    public ReadableByteChannel getPreparedTransaction( int identifier ) throws IOException
     {
         throw new UnsupportedOperationException();
     }
